requirements.txt
Django==4.0.1
django_versatileimagefield==2.2
drf_flex_fields==0.9.7
djangorestframework_simplejwt==5.0.0
djangorestframework==3.13.1
